+++ example/src/app/app.module.ts
@@ -2,7 +2,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 import { NgxTypeaheadModule } from 'ngx-typeahead';
-import { ArithmeticModule, SumService } from 'ngx-tester';
 
 import { AppComponent } from './app.component';
 
@@ -14,11 +13,9 @@ import { AppComponent } from './app.component';
   imports: [
     BrowserModule,
     HttpClientModule,
-    ArithmeticModule,
     NgxTypeaheadModule
   ],
   providers: [
-    SumService
   ],
   bootstrap: [AppComponent]
 })